Don't calculate the mount point on the fly
This stops the mount point from being calculated on the fly in various
classes. The mount point is now only calculated when the mntent is first
generated. A file system will now keep its mount point if the default
mount point directory is changed which keeps services from breaking.

The file system class also no longer takes a configuration object but
just a bunch of strings and the back end class.