Detect which mount file to use more reliably
Since OMV changed to mount file systems with a device path instead of
UUID= the code stopped working since '/proc/mounts' now contains slashes
in the fsname and the detection only checked that there shouldn't be
one.

The general rules for the new detection is that entries in
'/proc/mounts' doesn't start with a slash and doesn't necessarily have
any at all. Entries for mergerfs in '/etc/fstab' should always start
with a slash.